+++ Sources/StandardLibrary/Dictionary+Utils.swift
@@ -47,17 +47,6 @@ public extension Dictionary {
     return removedElements
   }
 
-  /// **Mechanica**
-  ///
-  /// Returns a random pair of (key,value) from `self`.
-  public func random() -> Element {
-    let index = Int.random(lowerBound: 0, upperBound: count)
-    let key = Array(keys)[index]
-    let value = self[key]!
-
-    return (key, value)
-  }
-
 }
 
 public extension Dictionary where Key: ExpressibleByStringLiteral {