+++ Sources/UIKit/UIButton+Utils.swift
@@ -34,7 +34,7 @@ extension UIButton {
   /// - Parameters:
   ///   - color: The background color to use for the specified state.
   ///   - state: The state that uses the specified image (defaults to normal.
-  public func setBackgroundColor(_ color: UIColor, for state: UIControlState = .normal) {
+  public func setBackgroundColor(_ color: UIColor, for state: UIControl.State = .normal) {
     let colorImage = UIImage(color: color, size: CGSize(width: 1, height: 1))
 
     setBackgroundImage(colorImage, for: state)